+++ src/TrelloDotNet.Tests/IntegrationTests/ActionTests.cs
@@ -1,4 +1,5 @@
 ﻿using TrelloDotNet.Model;
+using TrelloDotNet.Model.Options.GetActionsOptions;
 using TrelloDotNet.Model.Webhook;
 
 namespace TrelloDotNet.Tests.IntegrationTests;
@@ -16,7 +17,10 @@ public class ActionTests(TestFixtureWithNewBoard fixture) : TestBase, IClassFixt
         string newName = _board.Name + "GetActionsOfBoard";
         _board.Name = newName;
         await TrelloClient.UpdateBoardAsync(_board);
-        var actions = await TrelloClient.GetActionsOfBoardAsync(_boardId, [WebhookActionTypes.UpdateBoard]);
+        var actions = await TrelloClient.GetActionsOfBoardAsync(_boardId, new GetActionsOptions
+        {
+            Filter = [WebhookActionTypes.UpdateBoard]
+        });
         Assert.Contains(actions, x => x.Type == WebhookActionTypes.UpdateBoard && x.Data.Board.Name == newName && x.Data.Old.Name == nameBefore);
     }
 
@@ -27,7 +31,10 @@ public class ActionTests(TestFixtureWithNewBoard fixture) : TestBase, IClassFixt
         string newName = _organization.DisplayName + "GetActionsForOrganizations";
         _organization.DisplayName = newName;
         await TrelloClient.UpdateOrganizationAsync(_organization);
-        var actions = await TrelloClient.GetActionsForOrganizationsAsync(_organization.Id, [WebhookActionTypes.UpdateOrganization]);
+        var actions = await TrelloClient.GetActionsForOrganizationsAsync(_organization.Id, new GetActionsOptions
+        {
+            Filter = [WebhookActionTypes.UpdateOrganization]
+        });
         Assert.Contains(actions, x => x.Type == WebhookActionTypes.UpdateOrganization && x.Data.Organization.Name == newName && x.Data.Old.DisplayName == nameBefore);
     }
 
@@ -40,7 +47,10 @@ public class ActionTests(TestFixtureWithNewBoard fixture) : TestBase, IClassFixt
         const string newName = testName + "X";
         card.Name = newName;
         await TrelloClient.UpdateCardAsync(card);
-        var actions = await TrelloClient.GetActionsOnCardAsync(card.Id, [WebhookActionTypes.UpdateCard]);
+        var actions = await TrelloClient.GetActionsOnCardAsync(card.Id, new GetActionsOptions
+        {
+            Filter = [WebhookActionTypes.UpdateCard]
+        });
         Assert.Contains(actions, x => x.Type == WebhookActionTypes.UpdateCard && x.Data.Card.Id == card.Id && x.Data.Card.Name == newName && x.Data.Old.Name == testName);
     }
 
@@ -53,7 +63,10 @@ public class ActionTests(TestFixtureWithNewBoard fixture) : TestBase, IClassFixt
         const string newName = testName + "X";
         card.Name = newName;
         await TrelloClient.UpdateCardAsync(card);
-        var actions = await TrelloClient.GetActionsForListAsync(list.Id, [WebhookActionTypes.UpdateCard]);
+        var actions = await TrelloClient.GetActionsForListAsync(list.Id, new GetActionsOptions
+        {
+            Filter = [WebhookActionTypes.UpdateCard]
+        });
         Assert.Contains(actions, x => x.Type == WebhookActionTypes.UpdateCard && x.Data.Card.Id == card.Id && x.Data.Card.Name == newName && x.Data.Old.Name == testName);
     }
 
@@ -67,7 +80,10 @@ public class ActionTests(TestFixtureWithNewBoard fixture) : TestBase, IClassFixt
         card.Name = newName;
         await TrelloClient.UpdateCardAsync(card);
         Member member = await TrelloClient.GetTokenMemberAsync();
-        var actions = await TrelloClient.GetActionsForMemberAsync(member.Id, [WebhookActionTypes.UpdateCard]);
+        var actions = await TrelloClient.GetActionsForMemberAsync(member.Id, new GetActionsOptions
+        {
+            Filter = [WebhookActionTypes.UpdateCard]
+        });
         Assert.Contains(actions, x => x.Type == WebhookActionTypes.UpdateCard && x.Data.Card.Id == card.Id && x.Data.Card.Name == newName && x.Data.Old.Name == testName);
     }
 }