+++ src/TrelloDotNet/TrelloClient.Actions.cs
@@ -1,14 +1,28 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using TrelloDotNet.Control;
 using TrelloDotNet.Model;
 using TrelloDotNet.Model.Actions;
+using TrelloDotNet.Model.Options.GetActionsOptions;
 
 namespace TrelloDotNet
 {
     public partial class TrelloClient
     {
+        /// <summary>
+        /// Get the most recent Actions (Changelog Events) of a board
+        /// </summary>
+        /// <param name="boardId">The Id of the Board</param>
+        /// <param name="options">Options on how and what is retrieved</param>
+        /// <param name="cancellationToken">Cancellation Token</param>
+        /// <returns>List of most Recent Trello Actions</returns>
+        public async Task<List<TrelloAction>> GetActionsOfBoardAsync(string boardId, GetActionsOptions options, CancellationToken cancellationToken = default)
+        {
+            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsOnBoard(boardId), options, cancellationToken);
+        }
+
         /// <summary>
         /// Get the most recent Actions (Changelog Events) of a board
         /// </summary>
@@ -20,9 +34,29 @@ namespace TrelloDotNet
         /// <param name="before">An Action ID</param>
         /// <param name="since">An Action ID</param>
         /// <returns>List of most Recent Trello Actions</returns>
+        [Obsolete("Use overload the take in the 'GetActionsOptions' instead [Will be removed in v2.0 of this nuGet Package]")]
         public async Task<List<TrelloAction>> GetActionsOfBoardAsync(string boardId, List<string> filter = null, int limit = 50, CancellationToken cancellationToken = default, int page = 0, string before = null, string since = null)
         {
-            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsOnBoard(boardId), filter, limit, cancellationToken, page, before, since);
+            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsOnBoard(boardId), new GetActionsOptions
+            {
+                Limit = limit,
+                Before = before,
+                Filter = filter,
+                Page = page,
+                Since = since
+            }, cancellationToken);
+        }
+
+        /// <summary>
+        /// Get the most recent Actions (Changelog Events) on a Card
+        /// </summary>
+        /// <param name="cardId">The Id of the Card</param>
+        /// <param name="options">Options on how and what is retrieved</param>
+        /// <param name="cancellationToken">Cancellation Token</param>
+        /// <returns>List of most Recent Trello Actions</returns>
+        public async Task<List<TrelloAction>> GetActionsOnCardAsync(string cardId, GetActionsOptions options, CancellationToken cancellationToken = default)
+        {
+            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsOnCard(cardId), options, cancellationToken);
         }
 
         /// <summary>
@@ -36,9 +70,26 @@ namespace TrelloDotNet
         /// <param name="limit">How many recent events to get back; Default 50, Max 1000</param>
         /// <param name="cancellationToken">Cancellation Token</param>
         /// <returns>List of most Recent Trello Actions</returns>
+        [Obsolete("Use overload the take in the 'GetActionsOptions' instead [Will be removed in v2.0 of this nuGet Package]")]
         public async Task<List<TrelloAction>> GetActionsOnCardAsync(string cardId, List<string> filter = null, int limit = 50, CancellationToken cancellationToken = default)
         {
-            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsOnCard(cardId), filter, limit, cancellationToken);
+            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsOnCard(cardId), new GetActionsOptions
+            {
+                Limit = limit,
+                Filter = filter,
+            }, cancellationToken);
+        }
+
+        /// <summary>
+        /// Get the most recent Actions (Changelog Events) for a List
+        /// </summary>
+        /// <param name="listId">The Id of the List</param>
+        /// <param name="options">Options on how and what is retrieved</param>
+        /// <param name="cancellationToken">Cancellation Token</param>
+        /// <returns>List of most Recent Trello Actions</returns>
+        public async Task<List<TrelloAction>> GetActionsForListAsync(string listId, GetActionsOptions options, CancellationToken cancellationToken = default)
+        {
+            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsForList(listId), options, cancellationToken);
         }
 
         /// <summary>
@@ -49,9 +100,26 @@ namespace TrelloDotNet
         /// <param name="limit">How many recent events to get back; Default 50, Max 1000</param>
         /// <param name="cancellationToken">Cancellation Token</param>
         /// <returns>List of most Recent Trello Actions</returns>
+        [Obsolete("Use overload the take in the 'GetActionsOptions' instead [Will be removed in v2.0 of this nuGet Package]")]
         public async Task<List<TrelloAction>> GetActionsForListAsync(string listId, List<string> filter = null, int limit = 50, CancellationToken cancellationToken = default)
         {
-            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsForList(listId), filter, limit, cancellationToken);
+            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsForList(listId), new GetActionsOptions
+            {
+                Limit = limit,
+                Filter = filter,
+            }, cancellationToken);
+        }
+
+        /// <summary>
+        /// Get the most recent Actions (Changelog Events) for a Member
+        /// </summary>
+        /// <param name="memberId">The Id of the Member</param>
+        /// <param name="options">Options on how and what is retrieved</param>
+        /// <param name="cancellationToken">Cancellation Token</param>
+        /// <returns>List of most Recent Trello Actions</returns>
+        public async Task<List<TrelloAction>> GetActionsForMemberAsync(string memberId, GetActionsOptions options, CancellationToken cancellationToken = default)
+        {
+            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsForMember(memberId), options, cancellationToken);
         }
 
         /// <summary>
@@ -62,9 +130,29 @@ namespace TrelloDotNet
         /// <param name="limit">How many recent events to get back; Default 50, Max 1000</param>
         /// <param name="cancellationToken">Cancellation Token</param>
         /// <returns>List of most Recent Trello Actions</returns>
+        [Obsolete("Use overload the take in the 'GetActionsOptions' instead [Will be removed in v2.0 of this nuGet Package]")]
         public async Task<List<TrelloAction>> GetActionsForMemberAsync(string memberId, List<string> filter = null, int limit = 50, CancellationToken cancellationToken = default)
         {
-            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsForMember(memberId), filter, limit, cancellationToken);
+            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsForMember(memberId), new GetActionsOptions
+            {
+                Limit = limit,
+                Filter = filter,
+            }, cancellationToken);
+        }
+
+        /// <summary>
+        /// Get the most recent Actions (Changelog Events) for an Organization
+        /// </summary>
+        /// <remarks>
+        /// Only organization-specific actions will be returned. For the actions on the boards, see GetActionsOfBoardAsync
+        /// </remarks>
+        /// <param name="organizationId">The Id of the Organization</param>
+        /// <param name="options">Options on how and what is retrieved</param>
+        /// <param name="cancellationToken">Cancellation Token</param>
+        /// <returns>List of most Recent Trello Actions</returns>
+        public async Task<List<TrelloAction>> GetActionsForOrganizationsAsync(string organizationId, GetActionsOptions options, CancellationToken cancellationToken = default)
+        {
+            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsForOrganization(organizationId), options, cancellationToken);
         }
 
         /// <summary>
@@ -78,38 +166,45 @@ namespace TrelloDotNet
         /// <param name="limit">How many recent events to get back; Default 50, Max 1000</param>
         /// <param name="cancellationToken">Cancellation Token</param>
         /// <returns>List of most Recent Trello Actions</returns>
+        [Obsolete("Use overload the take in the 'GetActionsOptions' instead [Will be removed in v2.0 of this nuGet Package]")]
         public async Task<List<TrelloAction>> GetActionsForOrganizationsAsync(string organizationId, List<string> filter = null, int limit = 50, CancellationToken cancellationToken = default)
         {
-            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsForOrganization(organizationId), filter, limit, cancellationToken);
+            return await GetActionsFromSuffix(GetUrlBuilder.GetActionsForOrganization(organizationId), new GetActionsOptions
+            {
+                Limit = limit,
+                Filter = filter,
+            }, cancellationToken);
         }
 
-        private async Task<List<TrelloAction>> GetActionsFromSuffix(string suffix, List<string> filter, int limit, CancellationToken cancellationToken = default, int page = 0, string before = null, string since = null)
+        private async Task<List<TrelloAction>> GetActionsFromSuffix(string suffix, GetActionsOptions options, CancellationToken cancellationToken = default)
         {
             var parameters = new List<QueryParameter>();
-            if (filter != null)
+            if (options.Filter != null)
             {
-                parameters.Add(new QueryParameter("filter", string.Join(",", filter)));
+                parameters.Add(new QueryParameter("filter", string.Join(",", options.Filter)));
             }
 
-            if (limit > 0)
+            if (options.Limit > 0)
             {
-                parameters.Add(new QueryParameter("limit", limit));
+                parameters.Add(new QueryParameter("limit", options.Limit));
             }
 
-            if (page > 0)
+            if (options.Page > 0)
             {
-                parameters.Add(new QueryParameter("page", page));
+                parameters.Add(new QueryParameter("page", options.Page));
             }
 
-            if (before != null)
+            if (options.Before != null)
             {
-                parameters.Add(new QueryParameter("before", before));
+                parameters.Add(new QueryParameter("before", options.Before));
             }
 
-            if (since != null)
+            if (options.Since != null)
             {
-                parameters.Add(new QueryParameter("since", since));
+                parameters.Add(new QueryParameter("since", options.Since));
             }
+
+            parameters.AddRange(options.AdditionalParameters);
 
             return await _apiRequestController.Get<List<TrelloAction>>(suffix, cancellationToken, parameters.ToArray());
         }